+++ R/zzz_utils.R
@@ -1,7 +1,7 @@
 #NAMESPACE ----------
 
 #' @importFrom dplyr all_of any_of arrange as_tibble bind_cols bind_rows desc distinct everything filter group_by group_split mutate pull relocate rename rename_with select slice_max slice_min starts_with ungroup vars
-#' @importFrom ggplot2 %+replace% aes coord_cartesian element_rect facet_grid facet_wrap labeller geom_area geom_histogram geom_hline geom_line geom_point geom_rug geom_segment geom_vline ggplot label_both labs theme_bw scale_x_continuous scale_y_continuous scale_shape_manual scale_color_manual stat_function theme
+#' @importFrom ggplot2 %+replace% aes coord_cartesian element_rect facet_grid facet_wrap labeller geom_area geom_histogram geom_hline geom_line geom_point geom_rug geom_segment geom_vline ggplot label_both labs theme_bw scale_x_continuous scale_y_continuous scale_shape_manual scale_color_manual scale_linetype_manual stat_function theme
 #' @importFrom magrittr %>%
 #' @importFrom mrgsolve as.list data_set ev is.mrgmod mread mcode mrgsim mrgsim_df mrgsim_q mvgauss obsaug omat param realize_addl smat zero_re
 #' @importFrom optimx optimx
@@ -26,6 +26,19 @@ odiag <- function(x){
   diag(omat(x, make = T))
 }
 
+
+#' Get quantile from omega diag and a probability
+#'
+#' @param x model object
+#' @param .p percentile
+#'
+#' @return a vector of numeric
+#' @noRd
+get_quantile <- function(x, .p){
+  if(!is.mrgmod(x)) stop("the first argument to lowbounds must be a model object", call. = F)
+  map_dbl(sqrt(odiag(x)), qnorm, p = .p, mean = 0)
+}
+
 #' Internal "mapbayr" model examples
 #'
 #' @export
@@ -37,4 +50,9 @@ mbrlib <- function(){
 my_percent <- function(x){
   stopifnot(is.numeric(x))
   paste0(round(x * 100, 0), "%")
+}
+
+eta_from_opt <- function(x){
+  stopifnot(is.data.frame(x))
+  unlist(x[,grepl("ETA", names(x))])
 }